Change API implementation  (#168)
