+++ backend/src/index.ts
@@ -4,6 +4,7 @@ import { logger } from 'hono/logger';
 import { prettyJSON } from 'hono/pretty-json';
 import auth from './api/auth';
 import book from './api/book';
+import googleBooks from './api/google-books';
 import loan from './api/loan';
 import user from './api/user';
 
@@ -40,5 +41,6 @@ app.route('/books', book);
 app.route('/users', user);
 app.route('/auth', auth);
 app.route('/loans', loan);
+app.route('/googlebooks', googleBooks);
 
 export default app;
